move PostModelTest from functional_test/tests.py to board/tests.py
